refactor: 공통적으로 사용되는 UI 요소들 RoomForm 도메인으로 리팩토링
