+++ src/RoomForm/components/PasswordInput.tsx
@@ -1,15 +1,14 @@
 import { useFormContext } from 'react-hook-form';
-import { Inputs } from '@/RoomSetting/constants/form';
 import { PASSWORD } from '@/RoomForm/constants/literals';
-import { errorStyle, labelStyle } from './styles';
-import { PasswordInput } from '@/shared/Input';
+import { errorStyle } from '../constants/styles';
+import { PasswordInput as PasswordInputComponent } from '@/shared/Input';
 
-const PasswordSection = () => {
+const PasswordInput = () => {
   const {
     register,
     setValue,
     formState: { errors }
-  } = useFormContext<Inputs>();
+  } = useFormContext<{ password: string }>();
 
   const handleChangePassword = (e: React.ChangeEvent<HTMLInputElement>) => {
     const value = e.target.value.replace(/[^0-9]/g, '');
@@ -19,8 +18,7 @@ const PasswordSection = () => {
 
   return (
     <>
-      <label className={labelStyle}>비밀번호</label>
-      <PasswordInput
+      <PasswordInputComponent
         {...register('password')}
         placeholder="비워두시면 공개방이 됩니다"
         maxLength={PASSWORD.max}
@@ -33,4 +31,4 @@ const PasswordSection = () => {
   );
 };
 
-export default PasswordSection;
+export default PasswordInput;
