+++ src/RoomNew/index.ts
@@ -3,15 +3,3 @@ export { default as PasswordStep } from './steps/PasswordStep';
 export { default as RoutineStep } from './steps/RoutineStep';
 export { default as SummaryStep } from './steps/SummaryStep';
 export { default as TimeStep } from './steps/TimeStep';
-
-export {
-  ROOM_TYPES,
-  BIRD,
-  TIME_RANGE,
-  ROUTINE_COUNT,
-  ROUTINE_NAME,
-  ROOM_NAME,
-  PASSWORD,
-  USER_COUNT,
-  ANNOUNCEMENT
-} from '../RoomForm/constants/literals';
