+++ src/RoomNew/steps/RoutineStep/index.tsx
@@ -1,13 +1,45 @@
-import RoutineSection from './RoutineSection';
-import NameSection from './NameSection';
-import MemberSection from './MemberSection';
+import { useFormContext } from 'react-hook-form';
+import { ROOM_NAME } from '@/RoomForm/constants/literals';
+import { errorStyle } from '../../constants/styles';
+import { Inputs } from '../../constants/form';
+import { descriptionStyle } from '../../constants/styles';
+import { detailSectionStyle, headingStyle } from './styles';
+import { RoutineInputs, MemberInput } from '@/RoomForm';
+import { Input } from '@/shared/Input';
 
 const RoutineStep = () => {
+  const {
+    register,
+    formState: { errors }
+  } = useFormContext<Inputs>();
+
   return (
     <>
-      <RoutineSection />
-      <NameSection />
-      <MemberSection />
+      <section className={detailSectionStyle}>
+        <h2 className={headingStyle}>
+          같이 <b>어떤 루틴</b>을 해볼까요?
+        </h2>
+        <RoutineInputs />
+      </section>
+      <section className={detailSectionStyle}>
+        <h2 className={headingStyle}>
+          <b>방 이름</b>을 지어주세요.
+          <Input
+            {...register('title')}
+            maxLength={ROOM_NAME.max}
+            placeholder="짧고 명확하게!"
+          />
+          {errors.title && <p className={errorStyle}>{errors.title.message}</p>}
+        </h2>
+      </section>
+      <section className={detailSectionStyle}>
+        <h2 className={headingStyle}>
+          <b>최대 몇 명</b>의 친구들과
+          <p>함께하고 싶으세요?</p>
+          <p className={descriptionStyle}>(본인 포함)</p>
+        </h2>
+        <MemberInput />
+      </section>
     </>
   );
 };
